Remove datagen config and fix argument marshaling (#4791)
This adds unnecessary complexity

---------
